requirements.txt
Flask>=2.1.0
requests>=2.26.0
Pillow>=8.4.0